+++ MonteCarlo.h
@@ -2,14 +2,15 @@
 #define _MONTECARLO_H_
 
 #include "Random.h"
+#include "Constants.h"
 
 // Initialise each of the NUM_H + NUM_L matrices
 // randomly with elements in the range [-1-i, 1+i],
 // and calculate (and set) initialial action
 void Matrices_Initialisation(
     struct pcg32_random_t *rng,
-    float complex *Matrices,
-    float *action,
+    REAL complex *Matrices,
+    double *action,
     int NUM_H,
     int NUM_L
     );
@@ -19,20 +20,20 @@ void Matrices_Initialisation(
 // individually. The acceptance counter is update for each matrix
 void Get_Next_MCMC_Element(
     struct pcg32_random_t *rng,
-    float complex *Matrices,
-    float *action,
+    REAL complex *Matrices,
+    double *action,
     int *sigmaAB,
     int **sigmaABCD,
     int NUM_H,
     int NUM_L,
     int *acc,
-    float step_size
+    double step_size
     );
 
 // Tune the step size based on some acceptance rate measured
 void tune_step_size(
-    float accepance_rate, // acceptance rate so far
-    float* step_size      // reference to step_size to be tuned
+    double accepance_rate, // acceptance rate so far
+    double* step_size      // reference to step_size to be tuned
     );
 
 #endif
